Reset-MailPF
This script will help you to reset the mail enabled public folder object in all Exchange Server Versions. 
This will not remove the public folder it self, it is recreating the mail objects in Active Directory.
All of the smtp address assigned to the Public Folders will be set to the new objects again.
You have the ability to backup the List of all mail enabled public folders, to disable all of them and to enable all of them again and assign the smtp addresses.
The File should be saved in default PowerShell modules folder under a folder with the same name as the script.